Move App into component
